core: fft: Merge into one file
