+++ src/core/electrostatics_magnetostatics/fft-common.hpp
@@ -28,274 +28,5 @@
 
 #include <fftw3.h>
 
-/************************************************
- * data types
- ************************************************/
-
-/** Structure for performing a 1D FFT.
- *
- *  This includes the information about the redistribution of the 3D
- *  FFT *grid before the actual FFT.
- */
-struct fft_forw_plan {
-  /** plan direction: 0 = Forward FFT, 1 = Backward FFT. */
-  int dir;
-  /** row direction of that FFT. */
-  int row_dir;
-  /** permutations from normal coordinate system. */
-  int n_permute;
-  /** number of 1D FFTs. */
-  int n_ffts;
-  /** plan for fft. */
-  fftw_plan our_fftw_plan;
-
-    /** size of local mesh before communication. */
-  int old_mesh[3];
-  /** size of local mesh after communication, also used for actual FFT. */
-  int new_mesh[3];
-  /** lower left point of local FFT mesh in global FFT mesh coordinates. */
-  int start[3];
-  /** size of new mesh (number of mesh points). */
-  int new_size;
-
-  /** number of nodes which have to communicate with each other. */
-  int g_size;
-  /** group of nodes which have to communicate with each other. */
-  int *group;
-
-  /** packing function for send blocks. */
-  void (*pack_function)(double const *const, double *const, int const *,
-                        int const *, int const *, int);
-  /** Send block specification. 6 integers for each node: start[3], size[3]. */
-  int *send_block;
-  /** Send block communication sizes. */
-  int *send_size;
-  /** Recv block specification. 6 integers for each node: start[3], size[3]. */
-  int *recv_block;
-  /** Recv block communication sizes. */
-  int *recv_size;
-  /** size of send block elements. */
-  int element;
-};
-
-/** Additional information for backwards FFT.*/
-struct fft_back_plan {
-  /** plan direction. (e.g. fftw macro)*/
-  int dir;
-  /** plan for fft. */
-  fftw_plan our_fftw_plan;
-
-    /** packing function for send blocks. */
-  void (*pack_function)(double const *const, double *const, int const *,
-                        int const *, int const *, int);
-};
-
-/** Information about the three one dimensional FFTs and how the nodes
- *  have to communicate inbetween.
- *
- *  @note FFT numbering starts with 1 for technical reasons (because we have 4
- *        node grids, the index 0 is used for the real space charge assignment
- *        grid).
- */
-struct fft_data_struct {
-  /** Information for forward FFTs. */
-  fft_forw_plan plan[4];
-  /** Information for backward FFTs. */
-  fft_back_plan back[4];
-
-  /** Whether FFT is initialized or not. */
-  int init_tag;
-
-  /** Maximal size of the communication buffers. */
-  int max_comm_size;
-
-  /** Maximal local mesh size. */
-  int max_mesh_size;
-
-  /** send buffer. */
-  double *send_buf;
-  /** receive buffer. */
-  double *recv_buf;
-  /** Buffer for receive data. */
-  double *data_buf;
-};
-
-/************************************************
- * DEFINES
- ************************************************/
-
-/** @name MPI tags for FFT communication */
-/*@{*/
-/** Tag for communication in fft_init() */
-/** Tag for communication in forw_grid_comm() */
-#define REQ_FFT_FORW 301
-/** Tag for communication in back_grid_comm() */
-#define REQ_FFT_BACK 302
-/*@}*/
-
-/** Initialize FFT data structure. */
-void fft_pre_init(fft_data_struct *fft);
-
-/** This ugly function does the bookkeeping: which nodes have to
- *  communicate to each other, when you change the node grid.
- *  Changing the domain decomposition requires communication. This
- *  function finds (hopefully) the best way to do this. As input it
- *  needs the two grids (grid1, grid2) and a linear list (node_list1)
- *  with the node identities for grid1. The linear list (node_list2)
- *  for the second grid is calculated. For the communication group of
- *  the calling node it calculates a list (group) with the node
- *  identities and the positions (pos1, pos2) of that nodes in grid1
- *  and grid2. The return value is the size of the communication
- *  group. It gives -1 if the two grids do not fit to each other
- *  (grid1 and grid2 have to be component wise multiples of each
- *  other. see e.g. \ref calc_2d_grid in \ref grid.cpp for how to do
- *  this.).
- *
- * \param[in]  grid1       The node grid you start with.
- * \param[in]  grid2       The node grid you want to have.
- * \param[in]  node_list1  Linear node index list for grid1.
- * \param[out] node_list2  Linear node index list for grid2.
- * \param[out] group       Communication group (node identity list) for the
- *                         calling node.
- * \param[out] pos         Positions of the nodes in grid2
- * \param[out] my_pos      Position of this_node in grid2.
- * \return Size of the communication group.
- */
-int fft_find_comm_groups(Vector3i const &grid1, Vector3i const &grid2,
-                         int const *node_list1, int *node_list2, int *group,
-                         int *pos, int *my_pos);
-
-/** Calculate the local fft mesh.  Calculate the local mesh (loc_mesh)
- *  of a node at position (n_pos) in a node grid (n_grid) for a global
- *  mesh of size (mesh) and a mesh offset (mesh_off (in mesh units))
- *  and store also the first point (start) of the local mesh.
- *
- * \param[in]  n_pos    Position of the node in n_grid.
- * \param[in]  n_grid   node grid.
- * \param[in]  mesh     global mesh dimensions.
- * \param[in]  mesh_off global mesh offset (see \ref p3m_data_struct).
- * \param[out] loc_mesh local mesh dimension.
- * \param[out] start    first point of local mesh in global mesh.
- * \return Number of mesh points in local mesh.
- */
-int fft_calc_local_mesh(int const n_pos[3], int const n_grid[3],
-                        int const mesh[3], double const mesh_off[3],
-                        int loc_mesh[3], int start[3]);
-
-/** Calculate a send (or recv.) block for grid communication during a
- *  decomposition change.  Calculate the send block specification
- *  (block = lower left corner and upper right corner) which a node at
- *  position (pos1) in the actual node grid (grid1) has to send to
- *  another node at position (pos2) in the desired node grid
- *  (grid2). The global mesh, subject to communication, is specified
- *  via its size (mesh) and its mesh offset (mesh_off (in mesh
- *  units)).
- *
- *  For the calculation of a receive block you have to change the arguments in
- * the following way: <br> pos1  - position of receiving node in the desired
- * node grid. <br> grid1 - desired node grid. <br> pos2  - position of the node
- * you intend to receive the data from in the actual node grid. <br> grid2 -
- * actual node grid.  <br>
- *
- *  \param[in]  pos1     Position of send node in grid1.
- *  \param[in]  grid1    node grid 1.
- *  \param[in]  pos2     Position of recv node in grid2.
- *  \param[in]  grid2    node grid 2.
- *  \param[in]  mesh     global mesh dimensions.
- *  \param[in]  mesh_off global mesh offset (see \ref p3m_data_struct).
- *  \param[out] block    send block specification.
- *  \return Size of the send block.
- */
-int fft_calc_send_block(int const pos1[3], int const grid1[3],
-                        int const pos2[3], int const grid2[3],
-                        int const mesh[3], double const mesh_off[3],
-                        int block[6]);
-
-/** pack a block (size[3] starting at start[3]) of an input 3d-grid
- *  with dimension dim[3] into an output 3d-block with dimension size[3].
- *
- *    The block with dimensions (size[0], size[1], size[2]) is stored
- *    in 'row-major-order' or 'C-order', that means the first index is
- *    changing slowest when running through the linear array. The
- *    element (i0 (slow), i1 (mid), i2 (fast)) has the linear index
- *    li = i2 + size[2] * (i1 + (size[1]*i0))
- *
- *  \param[in]  in      pointer to input 3d-grid.
- *  \param[out] out     pointer to output 3d-grid (block).
- *  \param[in]  start   start index of the block in the in-grid.
- *  \param[in]  size    size of the block (=dimension of the out-grid).
- *  \param[in]  dim     size of the in-grid.
- *  \param[in]  element size of a grid element (e.g. 1 for Real, 2 for Complex).
- */
-void fft_pack_block(double const *in, double *out, int const start[3],
-                    int const size[3], int const dim[3], int element);
-
-/** pack a block with dimensions (size[0] * size[1] * aize[2]) starting
- *  at start[3] of an input 3d-grid with dimension dim[3] into an
- *  output 3d-grid with dimensions (size[2] * size[0] * size[1]) with
- *  a simultaneous one-fold permutation of the indices.
- *
- * The permutation is defined as:
- * slow_in -> fast_out, mid_in ->slow_out, fast_in -> mid_out
- *
- * An element (i0_in , i1_in , i2_in ) is then
- * (i0_out = i1_in-start[1], i1_out = i2_in-start[2], i2_out = i0_in-start[0])
- * and for the linear indices we have:                              <br> li_in =
- * i2_in + size[2] * (i1_in + (size[1]*i0_in))          <br> li_out = i2_out +
- * size[0] * (i1_out + (size[2]*i0_out))
- *
- * For index definition see \ref fft_pack_block.
- *
- *  \param[in]  in      pointer to input 3d-grid.
- *  \param[out] out     pointer to output 3d-grid (block).
- *  \param[in]  start   start index of the block in the in-grid.
- *  \param[in]  size    size of the block (=dimension of the out-grid).
- *  \param[in]  dim     size of the in-grid.
- *  \param[in]  element size of a grid element (e.g. 1 for Real, 2 for Complex).
- */
-void fft_pack_block_permute1(double const *in, double *out, int const start[3],
-                             int const size[3], int const dim[3], int element);
-
-/** pack a block with dimensions (size[0] * size[1] * aize[2]) starting
- *  at start[3] of an input 3d-grid with dimension dim[3] into an
- *  output 3d-grid with dimensions (size[2] * size[0] * size[1]), this
- *  is a simultaneous two-fold permutation of the indices.
- *
- * The permutation is defined as:
- * slow_in -> mid_out, mid_in ->fast_out, fast_in -> slow_out
- *
- * An element (i0_in , i1_in , i2_in ) is then
- * (i0_out = i2_in-start[2], i1_out = i0_in-start[0], i2_out = i1_in-start[1])
- * and for the linear indices we have:                              <br> li_in =
- * i2_in + size[2] * (i1_in + (size[1]*i0_in))          <br> li_out = i2_out +
- * size[0] * (i1_out + (size[2]*i0_out))
- *
- * For index definition see \ref fft_pack_block.
- *
- *  \param[in]  in      pointer to input 3d-grid.
- *  \param[out] out     pointer to output 3d-grid (block).
- *  \param[in]  start   start index of the block in the in-grid.
- *  \param[in]  size    size of the block (=dimension of the out-grid).
- *  \param[in]  dim     size of the in-grid.
- *  \param[in]  element size of a grid element (e.g. 1 for Real, 2 for Complex).
- */
-void fft_pack_block_permute2(double const *in, double *out, int const start[3],
-                             int const size[3], int const dim[3], int element);
-
-/** unpack a 3d-grid input block (size[3]) into an output 3d-grid
- *  with dimension dim[3] at start position start[3].
- *
- *  see also \ref fft_pack_block.
- *
- *  \param[in]  in      pointer to input 3d-grid.
- *  \param[out] out     pointer to output 3d-grid (block).
- *  \param[in]  start   start index of the block in the in-grid.
- *  \param[in]  size    size of the block (=dimension of the out-grid).
- *  \param[in]  dim     size of the in-grid.
- *  \param[in]  element size of a grid element (e.g. 1 for Real, 2 for Complex).
- */
-void fft_unpack_block(double const *in, double *out, int const start[3],
-                      int const size[3], int const dim[3], int element);
-
 #endif /* defined(P3M) || defined(DP3M) */
 #endif /* _FFT_COMMON_H */
