+++ src/core/electrostatics_magnetostatics/fft.hpp
@@ -45,11 +45,108 @@
 #include "config.hpp"
 #if defined(P3M) || defined(DP3M)
 
-#include "fft-common.hpp"
+#include <utils/Vector.hpp>
+
+#include <fftw3.h>
+
+/************************************************
+ * data types
+ ************************************************/
+
+/** Structure for performing a 1D FFT.
+ *
+ *  This includes the information about the redistribution of the 3D
+ *  FFT *grid before the actual FFT.
+ */
+struct fft_forw_plan {
+    /** plan direction: 0 = Forward FFT, 1 = Backward FFT. */
+    int dir;
+    /** row direction of that FFT. */
+    int row_dir;
+    /** permutations from normal coordinate system. */
+    int n_permute;
+    /** number of 1D FFTs. */
+    int n_ffts;
+    /** plan for fft. */
+    fftw_plan our_fftw_plan;
+
+    /** size of local mesh before communication. */
+    int old_mesh[3];
+    /** size of local mesh after communication, also used for actual FFT. */
+    int new_mesh[3];
+    /** lower left point of local FFT mesh in global FFT mesh coordinates. */
+    int start[3];
+    /** size of new mesh (number of mesh points). */
+    int new_size;
+
+    /** number of nodes which have to communicate with each other. */
+    int g_size;
+    /** group of nodes which have to communicate with each other. */
+    int *group;
+
+    /** packing function for send blocks. */
+    void (*pack_function)(double const *const, double *const, int const *,
+                          int const *, int const *, int);
+    /** Send block specification. 6 integers for each node: start[3], size[3]. */
+    int *send_block;
+    /** Send block communication sizes. */
+    int *send_size;
+    /** Recv block specification. 6 integers for each node: start[3], size[3]. */
+    int *recv_block;
+    /** Recv block communication sizes. */
+    int *recv_size;
+    /** size of send block elements. */
+    int element;
+};
+
+/** Additional information for backwards FFT.*/
+struct fft_back_plan {
+    /** plan direction. (e.g. fftw macro)*/
+    int dir;
+    /** plan for fft. */
+    fftw_plan our_fftw_plan;
+
+    /** packing function for send blocks. */
+    void (*pack_function)(double const *const, double *const, int const *,
+                          int const *, int const *, int);
+};
+
+/** Information about the three one dimensional FFTs and how the nodes
+ *  have to communicate inbetween.
+ *
+ *  @note FFT numbering starts with 1 for technical reasons (because we have 4
+ *        node grids, the index 0 is used for the real space charge assignment
+ *        grid).
+ */
+struct fft_data_struct {
+    /** Information for forward FFTs. */
+    fft_forw_plan plan[4];
+    /** Information for backward FFTs. */
+    fft_back_plan back[4];
+
+    /** Whether FFT is initialized or not. */
+    int init_tag;
+
+    /** Maximal size of the communication buffers. */
+    int max_comm_size;
+
+    /** Maximal local mesh size. */
+    int max_mesh_size;
+
+    /** send buffer. */
+    double *send_buf;
+    /** receive buffer. */
+    double *recv_buf;
+    /** Buffer for receive data. */
+    double *data_buf;
+};
 
 /** \name Exported Functions */
 /************************************************************/
 /*@{*/
+
+/** Initialize FFT data structure. */
+void fft_pre_init(fft_data_struct *fft);
 
 /** Initialize everything connected to the 3D-FFT.
  *
@@ -81,6 +178,40 @@ void fft_perform_forw(double *data, fft_data_struct &fft);
  *  \param         fft    FFT plan.
  */
 void fft_perform_back(double *data, bool check_complex, fft_data_struct &fft);
+
+/** pack a block (size[3] starting at start[3]) of an input 3d-grid
+ *  with dimension dim[3] into an output 3d-block with dimension size[3].
+ *
+ *    The block with dimensions (size[0], size[1], size[2]) is stored
+ *    in 'row-major-order' or 'C-order', that means the first index is
+ *    changing slowest when running through the linear array. The
+ *    element (i0 (slow), i1 (mid), i2 (fast)) has the linear index
+ *    li = i2 + size[2] * (i1 + (size[1]*i0))
+ *
+ *  \param[in]  in      pointer to input 3d-grid.
+ *  \param[out] out     pointer to output 3d-grid (block).
+ *  \param[in]  start   start index of the block in the in-grid.
+ *  \param[in]  size    size of the block (=dimension of the out-grid).
+ *  \param[in]  dim     size of the in-grid.
+ *  \param[in]  element size of a grid element (e.g. 1 for Real, 2 for Complex).
+ */
+void fft_pack_block(double const *in, double *out, int const start[3],
+                    int const size[3], int const dim[3], int element);
+
+/** unpack a 3d-grid input block (size[3]) into an output 3d-grid
+ *  with dimension dim[3] at start position start[3].
+ *
+ *  see also \ref fft_pack_block.
+ *
+ *  \param[in]  in      pointer to input 3d-grid.
+ *  \param[out] out     pointer to output 3d-grid (block).
+ *  \param[in]  start   start index of the block in the in-grid.
+ *  \param[in]  size    size of the block (=dimension of the out-grid).
+ *  \param[in]  dim     size of the in-grid.
+ *  \param[in]  element size of a grid element (e.g. 1 for Real, 2 for Complex).
+ */
+void fft_unpack_block(double const *in, double *out, int const start[3],
+                      int const size[3], int const dim[3], int element);
 
 /*@}*/
 #endif
